+++ crates/thetadatadx/src/fpss/mod.rs
@@ -89,10 +89,12 @@ pub mod framing;
 mod io_loop;
 pub mod protocol;
 pub mod ring;
+mod session;
 
 use self::events::IoCommand;
 pub use self::events::{FpssControl, FpssData, FpssEvent};
 use self::io_loop::{io_loop, ping_loop, wait_for_login, LoginResult};
+pub use self::session::{reconnect, reconnect_delay};
 
 use std::collections::HashMap;
 use std::sync::atomic::{AtomicBool, AtomicI32, Ordering};
@@ -109,7 +111,6 @@ use tdbe::types::enums::{RemoveReason, StreamMsgType};
 use self::framing::{write_frame, Frame};
 use self::protocol::{
     build_credentials_payload, build_subscribe_payload, Contract, SubscriptionKind,
-    RECONNECT_DELAY_MS, TOO_MANY_REQUESTS_DELAY_MS,
 };
 
 // ---------------------------------------------------------------------------
@@ -150,9 +151,9 @@ pub struct FpssClient {
     /// Monotonically increasing request ID counter.
     next_req_id: AtomicI32,
     /// Active per-contract subscriptions for reconnection.
-    active_subs: Mutex<Vec<(SubscriptionKind, Contract)>>,
+    pub(super) active_subs: Mutex<Vec<(SubscriptionKind, Contract)>>,
     /// Active full-type (firehose) subscriptions for reconnection.
-    active_full_subs: Mutex<Vec<(SubscriptionKind, tdbe::types::enums::SecType)>>,
+    pub(super) active_full_subs: Mutex<Vec<(SubscriptionKind, tdbe::types::enums::SecType)>>,
     /// Server-assigned contract ID mapping.
     contract_map: Arc<Mutex<HashMap<i32, Contract>>>,
     /// The server address we connected to.
@@ -764,7 +765,7 @@ impl FpssClient {
 
     /// Send a command to the I/O thread. Maps channel-send failure to a
     /// `Disconnected` FPSS error.
-    fn send_cmd(&self, cmd: IoCommand) -> Result<(), Error> {
+    pub(in crate::fpss) fn send_cmd(&self, cmd: IoCommand) -> Result<(), Error> {
         self.cmd_tx
             .lock()
             .unwrap_or_else(std::sync::PoisonError::into_inner)
@@ -797,126 +798,6 @@ impl Drop for FpssClient {
 // Reconnection helper
 // ---------------------------------------------------------------------------
 
-/// Reconnect an FPSS client after a disconnect.
-///
-/// # Behavior (from `FPSSClient.java`)
-///
-/// 1. Wait `delay_ms` before attempting reconnection
-/// 2. Establish a new TLS connection
-/// 3. Re-authenticate
-/// 4. Re-subscribe all previously active subscriptions with `req_id = -1`
-///
-/// On `TOO_MANY_REQUESTS`: wait 130 seconds before reconnecting.
-/// On `ACCOUNT_ALREADY_CONNECTED`: do NOT reconnect (permanent error).
-///
-/// Source: `FPSSClient.java` reconnection logic in the main loop.
-#[allow(clippy::too_many_arguments)] // Reason: reconnection requires all FPSS state (subs, config, credentials) in one call.
-#[allow(clippy::missing_errors_doc)] // Reason: internal function, doc is on the module-level reconnect docs above.
-pub fn reconnect<F>(
-    creds: &Credentials,
-    hosts: &[(String, u16)],
-    previous_subs: Vec<(SubscriptionKind, Contract)>,
-    previous_full_subs: Vec<(SubscriptionKind, tdbe::types::enums::SecType)>,
-    delay_ms: u64,
-    ring_size: usize,
-    flush_mode: FpssFlushMode,
-    policy: ReconnectPolicy,
-    derive_ohlcvc: bool,
-    handler: F,
-) -> Result<FpssClient, Error>
-where
-    F: FnMut(&FpssEvent) + Send + 'static,
-{
-    tracing::info!(delay_ms, "waiting before FPSS reconnection");
-    thread::sleep(Duration::from_millis(delay_ms));
-
-    let client = FpssClient::connect(
-        creds,
-        hosts,
-        ring_size,
-        flush_mode,
-        policy,
-        derive_ohlcvc,
-        handler,
-    )?;
-
-    // Re-subscribe all previous per-contract subscriptions with req_id = -1
-    // Source: FPSSClient.java -- reconnect logic uses req_id = -1 for re-subscriptions
-    for (kind, contract) in &previous_subs {
-        let payload = build_subscribe_payload(-1, contract);
-        let code = kind.subscribe_code();
-
-        client.send_cmd(IoCommand::WriteFrame { code, payload })?;
-
-        tracing::debug!(
-            kind = ?kind,
-            contract = %contract,
-            "re-subscribed after reconnect (req_id=-1)"
-        );
-    }
-
-    // Re-subscribe all previous full-type (firehose) subscriptions with req_id = -1
-    for (kind, sec_type) in &previous_full_subs {
-        let payload = protocol::build_full_type_subscribe_payload(-1, *sec_type);
-        let code = kind.subscribe_code();
-
-        client.send_cmd(IoCommand::WriteFrame { code, payload })?;
-
-        tracing::debug!(
-            kind = ?kind,
-            sec_type = ?sec_type,
-            "re-subscribed full-type after reconnect (req_id=-1)"
-        );
-    }
-
-    // Store the re-subscribed lists
-    {
-        let mut subs = client
-            .active_subs
-            .lock()
-            .unwrap_or_else(std::sync::PoisonError::into_inner);
-        *subs = previous_subs;
-    }
-    {
-        let mut subs = client
-            .active_full_subs
-            .lock()
-            .unwrap_or_else(std::sync::PoisonError::into_inner);
-        *subs = previous_full_subs;
-    }
-
-    Ok(client)
-}
-
-/// Determine the reconnect delay based on the disconnect reason.
-///
-/// Source: `FPSSClient.java` -- reconnect logic checks `RemoveReason` to decide delay.
-///
-/// # Intentional divergence from Java (see jvm-deviations.md)
-///
-/// Java only treats `AccountAlreadyConnected` (code 6) as a permanent error,
-/// retrying forever on invalid credentials — which burns rate limits and never
-/// succeeds. We treat all 7 credential/account error codes as permanent because
-/// no amount of retrying will fix bad credentials. This is a deliberate
-/// improvement over the Java behavior.
-#[must_use]
-pub fn reconnect_delay(reason: RemoveReason) -> Option<u64> {
-    match reason {
-        // Permanent errors -- no amount of reconnection will fix bad credentials.
-        // Java only checks AccountAlreadyConnected here; we extend this to all
-        // credential errors. See jvm-deviations.md "Permanent Disconnect".
-        RemoveReason::AccountAlreadyConnected
-        | RemoveReason::InvalidCredentials
-        | RemoveReason::InvalidLoginValues
-        | RemoveReason::InvalidLoginSize
-        | RemoveReason::FreeAccount
-        | RemoveReason::ServerUserDoesNotExist
-        | RemoveReason::InvalidCredentialsNullUser => None,
-        RemoveReason::TooManyRequests => Some(TOO_MANY_REQUESTS_DELAY_MS),
-        _ => Some(RECONNECT_DELAY_MS),
-    }
-}
-
 // ---------------------------------------------------------------------------
 // Tests
 // ---------------------------------------------------------------------------
@@ -925,36 +806,6 @@ pub fn reconnect_delay(reason: RemoveReason) -> Option<u64> {
 mod tests {
     use super::*;
     use tdbe::types::price::Price;
-
-    #[test]
-    fn reconnect_delay_permanent() {
-        // All credential / account errors are permanent -- no reconnect.
-        assert_eq!(reconnect_delay(RemoveReason::AccountAlreadyConnected), None);
-        assert_eq!(reconnect_delay(RemoveReason::InvalidCredentials), None);
-        assert_eq!(reconnect_delay(RemoveReason::InvalidLoginValues), None);
-        assert_eq!(reconnect_delay(RemoveReason::InvalidLoginSize), None);
-        assert_eq!(reconnect_delay(RemoveReason::FreeAccount), None);
-        assert_eq!(reconnect_delay(RemoveReason::ServerUserDoesNotExist), None);
-        assert_eq!(
-            reconnect_delay(RemoveReason::InvalidCredentialsNullUser),
-            None
-        );
-    }
-
-    #[test]
-    fn reconnect_delay_too_many_requests() {
-        assert_eq!(
-            reconnect_delay(RemoveReason::TooManyRequests),
-            Some(130_000)
-        );
-    }
-
-    #[test]
-    fn reconnect_delay_normal() {
-        assert_eq!(reconnect_delay(RemoveReason::ServerRestarting), Some(2_000));
-        assert_eq!(reconnect_delay(RemoveReason::Unspecified), Some(2_000));
-        assert_eq!(reconnect_delay(RemoveReason::TimedOut), Some(2_000));
-    }
 
     #[test]
     fn fpss_event_default_exists() {
